Footer changed to TS and slight refactor
